fix: disable faceki and web3auth
