+++ app/constants/browserstackTestAccounts.ts
@@ -3,7 +3,6 @@ export const browserstackTestAccounts = [
   'test-automation',
   'john-doe',
   'olive-5',
-  'marry-14',
   'mary-14',
   'bond-03',
   'rock-64',
